+++ include/graphics.h
@@ -4023,6 +4023,8 @@ extern const u32 gItemIcon_RageCandyBar[];
 extern const u32 gItemIconPalette_RageCandyBar[];
 extern const u32 gItemIcon_SlowpokeTail[];
 extern const u32 gItemIconPalette_SlowpokeTail[];
+extern const u32 gItemIcon_GBPlayer[];
+extern const u32 gItemIconPalette_GBPlayer[];
 
 extern const u32 gItemIcon_ReturnToFieldArrow[];
 extern const u32 gItemIconPalette_ReturnToFieldArrow[];