+++ src/battle_main.c
@@ -1881,7 +1881,7 @@ static void FreeRestoreBattleData(void)
     gScanlineEffect.state = 3;
     gMain.inBattle = 0;
     ZeroEnemyPartyMons();
-    m4aSongNumStop(SE_LOW_HEALTH);
+    m4aSongNumStop(SE_LOW_HEALTH, FALSE);
     FreeMonSpritesGfx();
     FreeBattleSpritesData();
     FreeBattleResources();
@@ -5229,7 +5229,7 @@ static void ReturnFromBattleToOverworld(void)
             SetRoamerInactive(&gEnemyParty[0]);
     }
 
-    m4aSongNumStop(SE_LOW_HEALTH);
+    m4aSongNumStop(SE_LOW_HEALTH, FALSE);
     SetMainCallback2(gMain.savedCallback);
 }