+++ src/credits.c
@@ -20,6 +20,8 @@
 #include "trig.h"
 #include "graphics.h"
 #include "pokedex.h"
+#include "constants/flags.h"
+#include "constants/vars.h"
 #include "event_data.h"
 #include "random.h"
 
@@ -445,7 +447,7 @@ void CB2_StartCreditsSequence(void)
     BeginNormalPaletteFade(PALETTES_ALL, 0, 16, 0, RGB_BLACK);
     EnableInterrupts(INTR_FLAG_VBLANK);
     SetVBlankCallback(VBlankCB_Credits);
-    m4aSongNumStart(MUS_CREDITS);
+    m4aSongNumStart(MUS_CREDITS, FlagGet(FLAG_SYS_GBS_ENABLED));
     SetMainCallback2(CB2_Credits);
     sUsedSpeedUp = FALSE;
     sCreditsData = AllocZeroed(sizeof(struct CreditsData));
@@ -683,7 +685,7 @@ static void Task_CreditsTheEnd6(u8 taskId)
             FadeOutBGM(8);
 
         if (gTasks[taskId].tDelay == 6840)
-            m4aSongNumStart(MUS_END);
+            m4aSongNumStart(MUS_END, FlagGet(FLAG_SYS_GBS_ENABLED));
 
         gTasks[taskId].tDelay--;
     }