+++ src/data/graphics/items.h
@@ -817,3 +817,6 @@ const u32 gItemIconPalette_RageCandyBar[] = INCBIN_U32("graphics/items/icon_pale
 
 const u32 gItemIcon_SlowpokeTail[] = INCBIN_U32("graphics/items/icons/slowpoke_tail.4bpp.lz");
 const u32 gItemIconPalette_SlowpokeTail[] = INCBIN_U32("graphics/items/icon_palettes/slowpoke_tail.gbapal.lz");
+
+const u32 gItemIcon_GBPlayer[] = INCBIN_U32("graphics/items/icons/gb_player.4bpp.lz");
+const u32 gItemIconPalette_GBPlayer[] = INCBIN_U32("graphics/items/icon_palettes/gb_player.gbapal.lz");