+++ src/main_menu.c
@@ -1,6 +1,7 @@
 #include "global.h"
 #include "trainer_pokemon_sprites.h"
 #include "bg.h"
+#include "constants/flags.h"
 #include "constants/rgb.h"
 #include "constants/songs.h"
 #include "constants/trainers.h"
@@ -211,7 +212,8 @@ static void Task_NewGameClockSetIntro5(u8);
 static void Task_NewGameClockSetIntro6(u8);
 static void Task_NewGameOakSpeech_Init(u8);
 static void Task_DisplayMainMenuInvalidActionError(u8);
-static void AddOakSpeechObjects(u8);
+static void NewGameOakSpeech_CreateWooperSprite(u8);
+static void NewGameOakSpeech_CreatePlatformSprites(u8);
 static void Task_NewGameOakSpeech_WaitToShowBirch(u8);
 static void NewGameOakSpeech_StartFadeInTarget1OutTarget2(u8, u8);
 static void NewGameOakSpeech_StartFadePlatformOut(u8, u8);
@@ -790,7 +792,7 @@ static void Task_MainMenuCheckSaveFile(u8 taskId)
     else if (IsBGMStopped()) // coming from title screen, waiting for music to fade
     {
         BeginNormalPaletteFade(0xFFFFFFFF, 0, 0x10, 0, 0xFFFF); // fade from white
-        m4aSongNumStart(MUS_MAIN_MENU);
+        m4aSongNumStart(MUS_MAIN_MENU, FlagGet(FLAG_SYS_GBS_ENABLED));
     }
     else    // egads, music is not faded yet!
     {
@@ -1603,6 +1605,8 @@ void Task_NewGameClockSetIntro1(u8 taskId)
         // moved from new_game.c so it doesn't change up the time on us unexpectedly after setting
         if (gSaveFileStatus == SAVE_STATUS_EMPTY || gSaveFileStatus == SAVE_STATUS_CORRUPT)
             RtcReset();
+        
+        FlagClear(FLAG_SYS_GBS_ENABLED);
 
         gTasks[taskId].data[0] = 15;
         gTasks[taskId].func = Task_NewGameClockSetIntro2;
@@ -1747,7 +1751,8 @@ static void Task_NewGameOakSpeech_Init(u8 taskId)
     LoadPalette(sOakSpeechBgPal, 0, 64);
     gPlttBufferUnfaded[0] = RGB_BLACK;
     gPlttBufferFaded[0] = RGB_BLACK;
-    AddOakSpeechObjects(taskId);
+    NewGameOakSpeech_CreateWooperSprite(taskId);
+    NewGameOakSpeech_CreatePlatformSprites(taskId);
     LoadOakIntroBigSprite(INTRO_OAK, 0);
     BeginNormalPaletteFade(0xFFFFFFFF, 4, 16, 0, 0);
     gTasks[taskId].tSlideOffset = 0;
@@ -2343,7 +2348,7 @@ static void CB2_NewGameOakSpeech_ReturnFromNamingScreen(void)
     ResetSpriteData();
     FreeAllSpritePalettes();
     ResetAllPicSprites();
-    AddOakSpeechObjects(taskId);
+    NewGameOakSpeech_CreatePlatformSprites(taskId);
     LoadOakIntroBigSprite(gSaveBlock2Ptr->playerGender, 0);
 
     for (i = 0; i < 3; i++)
@@ -2390,22 +2395,19 @@ static void SpriteCB_MovePlayerDownWhileShrinking(struct Sprite *sprite)
     sprite->data[0] = y;
 }
 
-static u8 NewGameOakSpeech_CreateWooperSprite(u8 a, u8 b)
+static void NewGameOakSpeech_CreateWooperSprite(u8 taskId)
 {
-    return CreatePicSprite2(SPECIES_WOOPER, SHINY_ODDS, 0, 1, a, b, 14, -1);
-}
-
-void AddOakSpeechObjects(u8 taskId)
-{
-    u8 i;
-    u8 wooperSprite;
-    u8 spriteId;
-
-    wooperSprite = NewGameOakSpeech_CreateWooperSprite(96, 96);
+    u8 wooperSprite = CreatePicSprite2(SPECIES_WOOPER, SHINY_ODDS, 0, 1, 96, 96, 14, -1);
     gSprites[wooperSprite].callback = SpriteCB_Null;
     gSprites[wooperSprite].oam.priority = 0;
     gSprites[wooperSprite].invisible = TRUE;
     gTasks[taskId].tWooperSpriteId = wooperSprite;
+}
+
+static void NewGameOakSpeech_CreatePlatformSprites(u8 taskId)
+{
+    u8 i;
+    u8 spriteId;
 
     LoadCompressedSpriteSheet(&sCompressedSpriteSheet_OakPlatform);
     LoadSpritePalette(&sSpritePalette_OakPlatform);