+++ src/sound.c
@@ -2,16 +2,19 @@
 #include "gba/m4a_internal.h"
 #include "sound.h"
 #include "battle.h"
+#include "event_data.h"
 #include "m4a.h"
 #include "main.h"
 #include "pokemon.h"
-#include "constants/songs.h"
 #include "task.h"
+#include "constants/flags.h"
+#include "constants/songs.h"
 
 struct Fanfare
 {
     u16 songNum;
     u16 duration;
+    u16 durationGBS;
 };
 
 // ewram
@@ -41,24 +44,25 @@ static void Task_DuckBGMForPokemonCry(u8 taskId);
 static void RestoreBGMVolumeAfterPokemonCry(void);
 
 static const struct Fanfare sFanfares[] = {
-    [FANFARE_LEVEL_UP]            = { MUS_LEVEL_UP,             80 },
-    [FANFARE_OBTAIN_ITEM]         = { MUS_OBTAIN_ITEM,         160 },
-    [FANFARE_EVOLVED]             = { MUS_EVOLVED,             220 },
-    [FANFARE_OBTAIN_TMHM]         = { MUS_OBTAIN_TMHM,         220 },
-    [FANFARE_HEAL]                = { MUS_HEAL,                160 },
-    [FANFARE_OBTAIN_BADGE]        = { MUS_OBTAIN_BADGE,        340 },
-    [FANFARE_MOVE_DELETED]        = { MUS_MOVE_DELETED,        180 },
-    [FANFARE_OBTAIN_EGG]          = { MUS_OBTAIN_EGG,          160 },
-    [FANFARE_PKMNCHANNEL_INTERLUDE] = { MUS_PKMNCHANNEL_INTERLUDE, 430 },
-    [FANFARE_SLOTS_JACKPOT]       = { MUS_SLOTS_JACKPOT,       250 },
-    [FANFARE_SLOTS_WIN]           = { MUS_SLOTS_WIN,           150 },
-    [FANFARE_TOO_BAD]             = { MUS_TOO_BAD,             160 },
-    [FANFARE_RG_POKE_FLUTE]       = { MUS_RG_POKE_FLUTE,       450 },
-    [FANFARE_RG_OBTAIN_KEY_ITEM]  = { MUS_RG_OBTAIN_KEY_ITEM,  170 },
-    [FANFARE_RG_DEX_RATING]       = { MUS_RG_DEX_RATING,       196 },
-    [FANFARE_OBTAIN_B_POINTS]     = { MUS_OBTAIN_B_POINTS,     313 },
-    [FANFARE_OBTAIN_SYMBOL]       = { MUS_OBTAIN_SYMBOL,       318 },
-    [FANFARE_REGISTER_MATCH_CALL] = { FANFARE_REGISTER_MATCH_CALL, 135 },
+    [FANFARE_LEVEL_UP]              = { MUS_LEVEL_UP,               80,  70 },
+    [FANFARE_OBTAIN_ITEM]           = { MUS_OBTAIN_ITEM,           160, 140 },
+    [FANFARE_EVOLVED]               = { MUS_EVOLVED,               220, 180 },
+    [FANFARE_OBTAIN_TMHM]           = { MUS_OBTAIN_TMHM,           220, 180 },
+    [FANFARE_HEAL]                  = { MUS_HEAL,                  160, 160 },
+    [FANFARE_OBTAIN_BADGE]          = { MUS_OBTAIN_BADGE,          340, 250 },
+    [FANFARE_MOVE_DELETED]          = { MUS_MOVE_DELETED,          180, 180 },
+    [FANFARE_OBTAIN_EGG]            = { MUS_OBTAIN_EGG,            160, 150 },
+    [FANFARE_PKMNCHANNEL_INTERLUDE] = { MUS_PKMNCHANNEL_INTERLUDE, 430, 410 },
+    [FANFARE_SLOTS_JACKPOT]         = { MUS_SLOTS_JACKPOT,         250, 250 },
+    [FANFARE_SLOTS_WIN]             = { MUS_SLOTS_WIN,             150, 150 },
+    [FANFARE_TOO_BAD]               = { MUS_TOO_BAD,               160, 160 },
+    [FANFARE_RG_POKE_FLUTE]         = { MUS_RG_POKE_FLUTE,         450, 450 },
+    [FANFARE_RG_OBTAIN_KEY_ITEM]    = { MUS_RG_OBTAIN_KEY_ITEM,    170, 170 },
+    [FANFARE_RG_DEX_RATING]         = { MUS_RG_DEX_RATING,         196, 196 },
+    [FANFARE_OBTAIN_B_POINTS]       = { MUS_OBTAIN_B_POINTS,       313, 313 },
+    [FANFARE_OBTAIN_SYMBOL]         = { MUS_OBTAIN_SYMBOL,         318, 318 },
+    [FANFARE_REGISTER_MATCH_CALL]   = { MUS_REGISTER_PHONE,        190, 180 },
+    [FANFARE_RG_CAUGHT_INTRO]       = { MUS_RG_CAUGHT_INTRO,       230, 170 },
 };
 
 #define CRY_VOLUME  120 // was 125 in R/S
@@ -188,10 +192,18 @@ bool8 IsNotWaitingForBGMStop(void)
 void PlayFanfareByFanfareNum(u8 fanfareNum)
 {
     u16 songNum;
+    bool32 isGBSEnabled = FlagGet(FLAG_SYS_GBS_ENABLED);
     m4aMPlayStop(&gMPlayInfo_BGM);
     songNum = sFanfares[fanfareNum].songNum;
-    sFanfareCounter = sFanfares[fanfareNum].duration;
-    m4aSongNumStart(songNum);
+    if (isGBSEnabled)
+    {
+        sFanfareCounter = sFanfares[fanfareNum].durationGBS;
+    }
+    else
+    {
+        sFanfareCounter = sFanfares[fanfareNum].duration;
+    }
+    m4aSongNumStart(songNum, isGBSEnabled);
 }
 
 bool8 WaitFanfare(bool8 stop)
@@ -206,7 +218,7 @@ bool8 WaitFanfare(bool8 stop)
         if (!stop)
             m4aMPlayContinue(&gMPlayInfo_BGM);
         else
-            m4aSongNumStart(MUS_DUMMY);
+            m4aSongNumStart(MUS_DUMMY, FALSE);
 
         return TRUE;
     }
@@ -215,13 +227,13 @@ bool8 WaitFanfare(bool8 stop)
 // Unused
 void StopFanfareByFanfareNum(u8 fanfareNum)
 {
-    m4aSongNumStop(sFanfares[fanfareNum].songNum);
+    m4aSongNumStop(sFanfares[fanfareNum].songNum, FlagGet(FLAG_SYS_GBS_ENABLED));
 }
 
 void PlayFanfare(u16 songNum)
 {
-    s32 i;
-    for (i = 0; (u32)i < ARRAY_COUNT(sFanfares); i++)
+    int i;
+    for (i = 0; i < ARRAY_COUNT(sFanfares); i++)
     {
         if (sFanfares[i].songNum == songNum)
         {
@@ -269,10 +281,10 @@ void FadeInNewBGM(u16 songNum, u8 speed)
         songNum = 0;
     if (songNum == MUS_NONE)
         songNum = 0;
-    m4aSongNumStart(songNum);
+    m4aSongNumStart(songNum, FlagGet(FLAG_SYS_GBS_ENABLED));
     m4aMPlayImmInit(&gMPlayInfo_BGM);
     m4aMPlayVolumeControl(&gMPlayInfo_BGM, 0xFFFF, 0);
-    m4aSongNumStop(songNum);
+    m4aSongNumStop(songNum, FlagGet(FLAG_SYS_GBS_ENABLED));
     m4aMPlayFadeIn(&gMPlayInfo_BGM, speed);
 }
 
@@ -566,17 +578,17 @@ void PlayBGM(u16 songNum)
         songNum = 0;
     if (songNum == MUS_NONE)
         songNum = 0;
-    m4aSongNumStart(songNum);
+    m4aSongNumStart(songNum, FlagGet(FLAG_SYS_GBS_ENABLED));
 }
 
 void PlaySE(u16 songNum)
 {
-    m4aSongNumStart(songNum);
+    m4aSongNumStart(songNum, FlagGet(FLAG_SYS_GBS_ENABLED));
 }
 
 void PlaySE12WithPanning(u16 songNum, s8 pan)
 {
-    m4aSongNumStart(songNum);
+    m4aSongNumStart(songNum, FlagGet(FLAG_SYS_GBS_ENABLED));
     m4aMPlayImmInit(&gMPlayInfo_SE1);
     m4aMPlayImmInit(&gMPlayInfo_SE2);
     m4aMPlayPanpotControl(&gMPlayInfo_SE1, 0xFFFF, pan);
@@ -585,14 +597,14 @@ void PlaySE12WithPanning(u16 songNum, s8 pan)
 
 void PlaySE1WithPanning(u16 songNum, s8 pan)
 {
-    m4aSongNumStart(songNum);
+    m4aSongNumStart(songNum, FlagGet(FLAG_SYS_GBS_ENABLED));
     m4aMPlayImmInit(&gMPlayInfo_SE1);
     m4aMPlayPanpotControl(&gMPlayInfo_SE1, 0xFFFF, pan);
 }
 
 void PlaySE2WithPanning(u16 songNum, s8 pan)
 {
-    m4aSongNumStart(songNum);
+    m4aSongNumStart(songNum, FlagGet(FLAG_SYS_GBS_ENABLED));
     m4aMPlayImmInit(&gMPlayInfo_SE2);
     m4aMPlayPanpotControl(&gMPlayInfo_SE2, 0xFFFF, pan);
 }